+++ apps/doc/src/routes/(rails)/play/+page.server.ts
@@ -10,12 +10,12 @@ export const load = async ({parent}) => {
 	let content = await pages.fetchMarkdowns(page)
 
 	if (!content?.length) {
-		throw error(404, {message: 'Not found'})
+		error(404, {message: 'Not found'})
 	}
 	content = content[0]
 
 	if (!content?.meta) {
-		throw error(404, {message: 'Not found'})
+		error(404, {message: 'Not found'})
 	}
 
 	return {