+++ apps/doc/src/routes/(rails)/play/learning/[slug]/+page.server.ts
@@ -2,15 +2,15 @@ import {error} from '@sveltejs/kit'
 import {actions as parentActions} from '../+page.server'
 
 export const load = async ({parent, params}) => {
-	let {learning} = await parent()
-	let meta = learning.find((s) => {
+	const {learning} = await parent()
+	const meta = learning.find((s) => {
 		return s.slug === params.slug
 	})
 
 	if (!meta) {
-		throw error(404, {message: 'Sketch not found'})
+		error(404, {message: 'Sketch not found'})
 	}
-	meta
+
 	return {meta, layout: 'steam'}
 }