+++ apps/doc/src/routes/api/blog/+server.ts
@@ -7,10 +7,11 @@ const page = 'blog'
 export async function GET({url}) {
 	const posts = blog.markdowns.filter(({meta}) => meta.status !== 'draft')
 
+	// TODO: This would be a call to an external API or database
 	const content = await pages.fetchMarkdowns(page)
 
 	if (!content?.meta) {
-		throw error(404, {message: 'Not found'})
+		error(404, {message: 'Not found'})
 	}
 
 	const data = {