+++ packages/media/.eslintrc.js
@@ -1,4 +1,0 @@
-import config from '@fat-fuzzy/config'
-const {eslint} = config
-
-export default eslint