+++ packages/playbook/playwright.config.ts
@@ -1,4 +1,16 @@
-import config from '@fat-fuzzy/config'
-const {playwright} = config
 
-export default playwright
+import playwright from '@fat-fuzzy/config/playwright'
+
+/** @type { import("@playwright/test").PlaywrightTestConfig } */
+const playwrightConfig = {
+	...playwright,
+	timeout: 1000,
+	projects: [
+		{
+			name: 'Desktop',
+			testMatch: /(.+\.)?(test|spec)\.[jt]s/,
+		},
+	],
+}
+
+export default playwrightConfig